+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+`1.2.0`
+- Changed Replace to return primary key int
+- Added handled exception code list to insert parameters
+- Removed Spam from checking Connect()
+
 `1.1.0`
 - Added Replace function
 - Added XML Summaries for all functions